backend/requirements.txt
fastapi
uvicorn
python-dotenv
pillow
transformers
torch
httpx
requests
